ci: unblock all 4 build-and-test jobs — clippy + stale tests + default target
CI has been red on every push since 2026-04-13 because the "Unit tests"
step fails early and masks everything that follows. This commit fixes the
full chain so all four OS matrix jobs (ubuntu, windows, macos-14,
macos-latest) run cleanly through Build → Unit tests → Clippy → E2E.

Root causes fixed:

- bare_cpdt_report_enables_full_mode: the weight-aware gate (PR #88/#90/
  #91) now requires weights for `--cpdt full`, but the test didn't supply
  any. Added a tiny safetensors fixture and `--weights <path>` to match
  the new decision-table contract, preserving the test's `Mode: full`
  stdout assertions.

- Clippy 1.95.0 churn (~60 new-lint hits in nsl-runtime and nsl-codegen):
  ran `cargo clippy --fix` for the mechanical ones, added crate-level
  allows for FFI-boundary lints (`not_unsafe_ptr_arg_deref`,
  `missing_safety_doc`), new-version doc-style lints
  (`doc_overindented_list_items`, `doc_lazy_continuation`), and the
  `manual_checked_ops` / `needless_range_loop` / `too_many_arguments` /
  `len_without_is_empty` families where the flagged code is intentional.
  `--allow-dirty` stripped two test-only imports it mistook for unused;
  both restored under `#[cfg(test)]` at the test module's top.

- Stale flash_attention test: `a23_projection_tile_sweep_uses_register_
  based_smem_addresses` was pinned to a fused `mad.lo.u32` pattern that
  was deliberately rewritten to `mul.lo.u32` + `add.u32` (per MEMORY.md:
  `mad.lo.u32` is rejected by CUDA at runtime on PTX ISA 7.0). Updated
  the assertions to match the current emission.

- Insta snapshots for fa_v2 kernel_full / phase_finalize and the PCA
  forward-kernel snapshot picked up legitimate upstream changes
  (finalize-phase cvt reordering, comment annotation) that were never
  committed. Accepted the pending `.snap.new` files.

- SMEM validation test was pinned to a 48 KB static budget; validator
  was updated to the 99 KB dynamic budget long ago. Renamed +
  retargeted to `smem_total_under_dynamic_budget_for_all_supported`.

- PTX-comment ASCII invariant: removed an em-dash from the CSHA A5 Wo
  stub comment (ptxas rejects UTF-8 in PTX source, per MEMORY.md).

- health_ffi integration tests share a global `Mutex<HealthCollector>`
  but CI doesn't pass `--test-threads=1`; added a file-local serial
  guard so the suite is parallel-safe under the default workspace
  command.

- ptxas-validation tests (a4/a5) were configured with
  `RopeStyle::HalfSplit` but `emit_rope_k_epilogue` only implements
  `Adjacent`. Switched to Adjacent so the tests exercise the supported
  path (they also gate on ptxas in PATH, so they skip on CI runners
  regardless — this just makes them pass locally under CUDA).

- `parse_gpu_sm_from_target` panicked on the default
  `CompileOptions::target = "cuda"`, breaking every `@flash_attention`
  e2e test (m27_paged_attention, m27_rope_gqa). Mapped `"cuda"` to
  sm_80 (Ampere) as the implied default.

- Semantic ownership walker collapsible-match + scope declare's
  Err-variant size lint (`result_large_err` allowed with rationale —
  no callers inspect the Err payload, boxing would add cost on the
  hot success path for no win).

Local verification: all four CI steps pass on this branch — Build
finishes, Unit tests run 1650+ tests with 0 failures, Clippy emits
nothing, E2E runs 80 tests 0 failures.

+++ crates/nsl-codegen/src/cfie.rs
@@ -225,7 +225,7 @@ pub fn run(input: CfieInput) -> CfiePlan {
     // Kernel-launch accounting.
     let baseline = 500u32; // paper §1's "~500-1000 launches/token" figure
     let saved_persistent = (persistent.baseline_launches_per_layer
-        .saturating_sub(persistent.persistent_launches_per_layer)) as u32
+        .saturating_sub(persistent.persistent_launches_per_layer))
         * input.persistent_model.n_layers;
     // Sampling + residual layers: one fused kernel versus six.
     let saved_sampling = if sampling.is_fused() { 5 } else { 0 };
